+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!-- RETAIN THIS COMMENT. An entry template for a new version is automatically added each time `Invoke-psake version` is called. Fill in changes afterwards. -->
 
+* **v1.1** (2020-08-29):
+  * [enhancement] `ie` now also accepts and resolves *aliases* of external executables.
+  * [enhancement] `ie` now calls `Rscript.exe` with the required `\"`-escaping in Windows PowerShell (in PowerShell Core it always did).
+
 * **v1.0.10** (2020-08-21):
   * [dev] Improved tests, streamlined implementation.